refactor: 도메인이 포함된 함수를 domain 폴더로 이동
